Allow widget placeholder replacement on primitives
